requirements.txt
requests==2.31.0
python-dotenv==1.0.0
google-cloud-bigquery==3.15.0